+++ src/app/cart-list/cart-list.module.ts
@@ -1,5 +1,4 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
 
 import { SharedModule } from '../shared/shared.module';
 
@@ -8,8 +7,6 @@ import { CartItemComponent } from './components/cart-item/cart-item.component';
 
 @NgModule({
 	imports: [
-		CommonModule,
-
 		SharedModule
 	],
 	declarations: [CartListComponent, CartItemComponent],