+++ src/app/first/first.module.ts
@@ -1,5 +1,4 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
 
 import { SharedModule } from '../shared/shared.module';
 
@@ -7,8 +6,6 @@ import { FirstComponent } from './first.component';
 
 @NgModule({
 	imports: [
-		CommonModule,
-
 		SharedModule
 	],
 	declarations: [FirstComponent],