+++ src/app/shared/shared.module.ts
@@ -1,17 +1,25 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+
 import { AngularMaterialSharedModule } from './angular-material/angular-material-shared.module';
 
 import { HighlightDirective, DrawBorderDirective } from './directives';
+import { OrderByPipe } from './pipes';
 
 @NgModule({
-	declarations: [HighlightDirective, DrawBorderDirective],
+	declarations: [HighlightDirective, DrawBorderDirective, OrderByPipe],
 	imports: [CommonModule],
 	exports: [
+		CommonModule,
+		FormsModule,
+		ReactiveFormsModule,
+
+		AngularMaterialSharedModule,
+
 		HighlightDirective,
 		DrawBorderDirective,
-
-		AngularMaterialSharedModule
+		OrderByPipe
 	]
 })
 export class SharedModule { }